backend/requirements.txt
# 后端依赖
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
alembic==1.14.*
pydantic-settings==2.7.*
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
redis==5.2.*
httpx==0.28.*
websockets>=12.0,<14.0
aiosqlite==0.22.*
socksio==1.0.*
python-binance==1.0.*
pandas==2.2.*
# ta-lib==0.5.*  # 暂时注释，需要复杂的编译环境
numpy==2.1.*

# 开发依赖
pytest==8.3.*
pytest-asyncio==0.24.*
pytest-cov==6.0.*
ruff==0.8.*
mypy==1.13.*
